Changed conf file to remove the .* from the whitelist for extensions.  Don't really need it at this point since I'm not doing a wildcard matching.
Fixed a bug in ext matching, now properly matches the extensions in the ext whitelist.  Forgot to split the extList into a list which can be iterated.  Python is funny and flexible so it's easy not to have the string methods explode on you and instead return a very unexpected results.  Helps to read docs.